Get negative samples for combo from WiMCor
Till now, we used samples from AID as negative samples.
However AIDA samples consist of metonymic mentions as well,
which makes the combo dataset noisy.

We have now added a facility to extract negative samples
from WiMCor.

+++ code/preprocessing/prepro_combo.py
@@ -8,6 +8,9 @@ from preprocessing.prepro_aida import split_dev_test
 from preprocessing.prepro_aida import write_to_file
 from preprocessing.prepro_wimcor import process_wimcor
 
+LITERAL = 'LIT'
+METONYMIC = 'MET'
+
 def _parse_args():
     parser = argparse.ArgumentParser()
     parser.add_argument("--aida_folder",
@@ -20,6 +23,9 @@ def _parse_args():
                         type=float,
                         default=0.7,
                         help="ratio for split for train; rest is equally divided into dev and test")
+    parser.add_argument("--negative_source",
+                        default='wimcor',
+                        help="source for negative samples: aida or wimcor")
     parser.add_argument("--add_noise",
                         dest='add_noise',
                         action='store_true',
@@ -63,37 +69,50 @@ if __name__ == "__main__":
     print(args)
     create_necessary_folders(args.output_folder)
 
-    # get AIDA samples
-    split_dev_test(args.aida_folder+"testa_testb_aggregate_original", args.output_folder)
-    aida_samples = process_aida(args.output_folder+"temp_aida_dev",
-                                args.add_noise, args.noise_type,
-                                metotype='LIT',
-                                )
-    os.remove(args.output_folder + "temp_aida_dev")
-    os.remove(args.output_folder + "temp_aida_test")
+    # get negative samples
+    if args.negative_source == 'aida':
+        split_dev_test(args.aida_folder+"testa_testb_aggregate_original", args.output_folder)
+        negative_samples = process_aida(args.output_folder+"temp_aida_dev",
+                                        args.add_noise, args.noise_type,
+                                        metotype='LIT',
+                                        )
+        os.remove(args.output_folder + "temp_aida_dev")
+        os.remove(args.output_folder + "temp_aida_test")
+    elif args.negative_source == 'wimcor':
+        negative_samples, negative_ent_id_indices = process_wimcor(args.wimcor_folder+"wimcor_negative.xml",
+                                                                   args.add_noise, args.noise_type,
+                                                                   metotype=LITERAL,)
 
-    # get WiMCor samples
-    wimcor_samples, ent_id_indices = process_wimcor(args.wimcor_folder+"wimcor_positive.xml",
-                                                    args.add_noise, args.noise_type,
-                                                    metotype='MET',)
+        # reorder samples according to entity ids
+        if args.reorder_samples:
+            reordered_negative_samples = []
+            for ent_id in negative_ent_id_indices.keys():
+                for idx in negative_ent_id_indices[ent_id]:
+                    reordered_negative_samples.append(negative_samples[idx])
+            negative_samples = reordered_negative_samples
 
-    # reorder WiMCor samples according to entity ids
+    # get positive samples
+    positive_samples, positive_ent_id_indices = process_wimcor(args.wimcor_folder+"wimcor_positive.xml",
+                                                               args.add_noise, args.noise_type,
+                                                               metotype=METONYMIC,)
+
+    # reorder samples according to entity ids
     if args.reorder_samples:
-        reordered_wimcor_samples = []
-        for ent_id in ent_id_indices.keys():
-            for idx in ent_id_indices[ent_id]:
-                reordered_wimcor_samples.append(wimcor_samples[idx])
-        wimcor_samples = reordered_wimcor_samples
+        reordered_positive_samples = []
+        for ent_id in positive_ent_id_indices.keys():
+            for idx in positive_ent_id_indices[ent_id]:
+                reordered_positive_samples.append(positive_samples[idx])
+        positive_samples = reordered_positive_samples
 
     # split the data into train, dev and test partitions
     if args.strict_reordering:
         # unseen entities in dev and test; harder version
-        combo_samples_train = aida_samples[:int(args.split_ratio*len(aida_samples))] + \
-                                wimcor_samples[:int(args.split_ratio*len(wimcor_samples))]
-        combo_samples_dev = aida_samples[int(args.split_ratio*len(aida_samples)):int(args.split_ratio*len(aida_samples))+int((1-args.split_ratio)*len(aida_samples)//2)] + \
-                                wimcor_samples[int(args.split_ratio*len(wimcor_samples)):int(args.split_ratio*len(wimcor_samples))+int((1-args.split_ratio)*len(wimcor_samples)//2)]
-        combo_samples_test = aida_samples[int(args.split_ratio*len(aida_samples))+int((1-args.split_ratio)*len(aida_samples)//2):] + \
-                                wimcor_samples[int(args.split_ratio*len(wimcor_samples))+int((1-args.split_ratio)*len(wimcor_samples)//2):]
+        combo_samples_train = negative_samples[:int(args.split_ratio*len(negative_samples))] + \
+                                positive_samples[:int(args.split_ratio*len(positive_samples))]
+        combo_samples_dev = negative_samples[int(args.split_ratio*len(negative_samples)):int(args.split_ratio*len(negative_samples))+int((1-args.split_ratio)*len(negative_samples)//2)] + \
+                                positive_samples[int(args.split_ratio*len(positive_samples)):int(args.split_ratio*len(positive_samples))+int((1-args.split_ratio)*len(positive_samples)//2)]
+        combo_samples_test = negative_samples[int(args.split_ratio*len(negative_samples))+int((1-args.split_ratio)*len(negative_samples)//2):] + \
+                                positive_samples[int(args.split_ratio*len(positive_samples))+int((1-args.split_ratio)*len(positive_samples)//2):]
 
         # shuffle the data; use seed to ensure reproducibility
         random.seed(42)
@@ -102,10 +121,10 @@ if __name__ == "__main__":
         random.shuffle(combo_samples_test)
     else:
         # similar entities in dev and test;
-        combo_samples_train = aida_samples[:int(args.split_ratio*len(aida_samples))] + \
-                                wimcor_samples[:int(args.split_ratio*len(wimcor_samples))]
-        combo_samples_rest = aida_samples[int(args.split_ratio*len(aida_samples)):] + \
-                                wimcor_samples[int(args.split_ratio*len(wimcor_samples)):]
+        combo_samples_train = negative_samples[:int(args.split_ratio*len(negative_samples))] + \
+                                positive_samples[:int(args.split_ratio*len(positive_samples))]
+        combo_samples_rest = negative_samples[int(args.split_ratio*len(negative_samples)):] + \
+                                positive_samples[int(args.split_ratio*len(positive_samples)):]
 
             # shuffle the data; use seed to ensure reproducibility
         random.seed(42)
@@ -127,32 +146,37 @@ if __name__ == "__main__":
     elif not args.reorder_samples and not args.strict_reordering:
         is_reordered = ''
 
+    if args.negative_source == 'aida':
+        is_extended = ''
+    elif args.negative_source == 'wimcor':
+        is_extended = '_extended'
+
     if args.add_noise and args.noise_type=='distort_meto_labels':
         write_to_file(combo_samples_train,
-                      args.output_folder+"combo"+is_reordered+"_metolabelsdistorted"+"_train.txt")
+                      args.output_folder+"combo"+is_extended+is_reordered+"_metolabelsdistorted"+"_train.txt")
         write_to_file(combo_samples_dev,
-                      args.output_folder+"combo"+is_reordered+"_metolabelsdistorted"+"_dev.txt")
+                      args.output_folder+"combo"+is_extended+is_reordered+"_metolabelsdistorted"+"_dev.txt")
         write_to_file(combo_samples_test,
-                      args.output_folder+"combo"+is_reordered+"_metolabelsdistorted"+"_test.txt")
+                      args.output_folder+"combo"+is_extended+is_reordered+"_metolabelsdistorted"+"_test.txt")
     elif args.add_noise and args.noise_type=='distort_el_labels':
         write_to_file(combo_samples_train,
-                      args.output_folder+"combo"+is_reordered+"_ellabelsdistorted"+"_train.txt")
+                      args.output_folder+"combo"+is_extended+is_reordered+"_ellabelsdistorted"+"_train.txt")
         write_to_file(combo_samples_dev,
-                      args.output_folder+"combo"+is_reordered+"_ellabelsdistorted"+"_dev.txt")
+                      args.output_folder+"combo"+is_extended+is_reordered+"_ellabelsdistorted"+"_dev.txt")
         write_to_file(combo_samples_test,
-                      args.output_folder+"combo"+is_reordered+"_ellabelsdistorted"+"_test.txt")
+                      args.output_folder+"combo"+is_extended+is_reordered+"_ellabelsdistorted"+"_test.txt")
     elif args.add_noise and args.noise_type=='distort_context':
         write_to_file(combo_samples_train,
-                      args.output_folder+"combo"+is_reordered+"_contextdistorted"+"_train.txt")
+                      args.output_folder+"combo"+is_extended+is_reordered+"_contextdistorted"+"_train.txt")
         write_to_file(combo_samples_dev,
-                      args.output_folder+"combo"+is_reordered+"_contextdistorted"+"_dev.txt")
+                      args.output_folder+"combo"+is_extended+is_reordered+"_contextdistorted"+"_dev.txt")
         write_to_file(combo_samples_test,
-                      args.output_folder+"combo"+is_reordered+"_contextdistorted"+"_test.txt")
+                      args.output_folder+"combo"+is_extended+is_reordered+"_contextdistorted"+"_test.txt")
     elif not args.add_noise:
         # no perturbation; data as is;
         write_to_file(combo_samples_train,
-                      args.output_folder+"combo"+is_reordered+"_train.txt")
+                      args.output_folder+"combo"+is_extended+is_reordered+"_train.txt")
         write_to_file(combo_samples_dev,
-                      args.output_folder+"combo"+is_reordered+"_dev.txt")
+                      args.output_folder+"combo"+is_extended+is_reordered+"_dev.txt")
         write_to_file(combo_samples_test,
-                      args.output_folder+"combo"+is_reordered+"_test.txt")
+                      args.output_folder+"combo"+is_extended+is_reordered+"_test.txt")
